improves README to better explain column assignment, renames coordRect to column2CoordRect
