+++ src/arquint/coord/column2CoordRect.js
